Updates to recon ITs
Inline all plots into the test classes.

Convert the 2015 tests to use the new ReconTest class.

Additional cleanup and reorganization.
