+++ integration-tests/src/test/java/org/hps/test/it/EngRun2015MollerReconTest.java
@@ -2,90 +2,332 @@ package org.hps.test.it;
 
 import static java.lang.Math.abs;
 
-import java.io.File;
-import java.io.IOException;
+import java.util.List;
 
-import org.hps.evio.EvioToLcio;
-import org.hps.util.test.TestUtil;
-import org.hps.util.test.TestOutputFile;
-import org.lcsim.util.aida.AIDA;
-import org.lcsim.util.loop.LCSimLoop;
+import org.hps.recon.ecal.cluster.ClusterUtilities;
+import org.hps.recon.tracking.TrackType;
+import org.lcsim.event.Cluster;
+import org.lcsim.event.EventHeader;
+import org.lcsim.event.ReconstructedParticle;
+import org.lcsim.event.Track;
+import org.lcsim.event.Vertex;
+import org.lcsim.geometry.Detector;
+import org.lcsim.math.chisq.ChisqProb;
 
-import hep.aida.IAnalysisFactory;
 import hep.aida.IHistogram1D;
-import hep.aida.ITree;
-import junit.framework.TestCase;
+import hep.aida.IHistogram2D;
+import hep.physics.vec.BasicHep3Matrix;
+import hep.physics.vec.Hep3Vector;
+import hep.physics.vec.VecOp;
 
 /**
- *
- * @author Norman A. Graf
+ * Test Moller reconstruction for 2015 engineering run
  */
-public class EngRun2015MollerReconTest extends TestCase {
+public class EngRun2015MollerReconTest extends ReconTest {
 
-    static final String testFileName = "hps_005772_mollerskim_10k.evio";
-    static final String fieldmapName = "HPS-EngRun2015-Nominal-v6-0-fieldmap_v3";
-    static final String steeringFileName = "/org/hps/steering/recon/legacy_drivers/EngineeringRun2015FullRecon.lcsim";
-    private final int nEvents = 1000;
-    private String aidaOutputFile = "target/test-output/EngRun2015MollerReconTest/EngRun2015MollerReconTest";
+    static final String DETECTOR = "HPS-EngRun2015-Nominal-v6-0-fieldmap_v3";
+    static final String TEST_FILE_NAME = "hps_005772_mollerskim_10k.evio";
+    static final String STEERING = "/org/hps/steering/recon/legacy_drivers/EngineeringRun2015FullRecon.lcsim";
+    static final int NEVENTS = 5000;
+    static final long MAX_EVENT_TIME = -1;
 
-    public void testIt() throws Exception {
-        File evioInputFile = TestUtil.downloadTestFile(testFileName);
-        File outputFile = new TestOutputFile(EngRun2015MollerReconTest.class, "recon");
-        String args[] = {"-r", "-x", steeringFileName, "-d",
-            fieldmapName, "-D", "outputFile=" + outputFile.getPath(), "-n", String.format("%d", nEvents),
-            evioInputFile.getPath()};
-        EvioToLcio.main(args);
-        LCSimLoop loop = new LCSimLoop();
-        EngRun2015MollerRecon reconDriver = new EngRun2015MollerRecon();
-        aidaOutputFile = new TestOutputFile(getClass().getSimpleName()).getPath() + File.separator + this.getClass().getSimpleName();
-        reconDriver.setAidaFileName(aidaOutputFile);
-        loop.add(reconDriver);
-        try {
-            loop.setLCIORecordSource(new File(outputFile.getPath() + ".slcio"));
-            loop.loop(-1);
-        } catch (IOException e) {
-            throw new RuntimeException(e);
-        }
-        comparePlots();
+    public EngRun2015MollerReconTest() {
+        super(EngRun2015MollerReconTest.class,
+                DETECTOR,
+                TEST_FILE_NAME,
+                STEERING,
+                NEVENTS,
+                new PlotDriver(),
+                DEFAULT_TOLERANCE,
+                MAX_EVENT_TIME,
+                true,
+                true);
     }
 
-    public void comparePlots() throws Exception {
-        AIDA aida = AIDA.defaultInstance();
-        final IAnalysisFactory af = aida.analysisFactory();
+    public static class PlotDriver extends RefDriver {
 
-        File aidaRefFile = TestUtil.downloadRefPlots("EngRun2015MollerReconTest");
+        String[] vertexCollectionNames = {"UnconstrainedMollerVertices", "BeamspotConstrainedMollerVertices", "TargetConstrainedMollerVertices"};
+        private final BasicHep3Matrix beamAxisRotation = new BasicHep3Matrix();
 
-        File aidaTstFile = new File(aidaOutputFile+".aida");
+        private Double _beamEnergy = 1.056;
+        private double _psumDelta = 0.06;
+        private double _thetasumCut = 0.07;
+        private double _trackChi2NdfCut = 8.; //corresponds to chisquared cut of 40 for 5-hit tracks
+        private double psumMin = (1 - _psumDelta) * _beamEnergy;
+        private double psumMax = (1 + _psumDelta) * _beamEnergy;
 
-        ITree ref = af.createTreeFactory().create(aidaRefFile.getAbsolutePath());
-        ITree tst = af.createTreeFactory().create(aidaTstFile.getAbsolutePath());
+        private IHistogram1D invMassHist_UnconstrainedMollerVertices = aida.histogram1D("UnconstrainedMollerVertices/Moller Invariant Mass", 200, 0., 0.1);
+        private IHistogram1D pHist_UnconstrainedMollerVertices = aida.histogram1D("UnconstrainedMollerVertices/Moller Momentum", 200, 0., 3.0);
+        private IHistogram1D pxHist_UnconstrainedMollerVertices = aida.histogram1D("UnconstrainedMollerVertices/Moller x Momentum", 200, -0.01, 0.01);
+        private IHistogram1D pyHist_UnconstrainedMollerVertices = aida.histogram1D("UnconstrainedMollerVertices/Moller y Momentum", 200, -0.01, 0.01);
+        private IHistogram1D pzHist_UnconstrainedMollerVertices = aida.histogram1D("UnconstrainedMollerVertices/Moller z Momentum", 200, 0., 3.0);
+        private IHistogram1D trkpHist_UnconstrainedMollerVertices = aida.histogram1D("UnconstrainedMollerVertices/Moller Track Momentum", 100, 0.25, 1.75);
+        private IHistogram1D trkptopHist_UnconstrainedMollerVertices = aida.histogram1D("UnconstrainedMollerVertices/Moller Top Track Momentum", 100, 0.25, 1.75);
+        private IHistogram1D trkpbotHist_UnconstrainedMollerVertices = aida.histogram1D("UnconstrainedMollerVertices/Moller Bottom Track Momentum", 100, 0.25, 1.75);
+        private IHistogram1D trkNhitsHist_UnconstrainedMollerVertices = aida.histogram1D("UnconstrainedMollerVertices/Moller Track Number of Hits", 7, -0.5, 6.5);
+        private IHistogram1D trkChisqHist_UnconstrainedMollerVertices = aida.histogram1D("UnconstrainedMollerVertices/Moller Track Chisq per DoF", 100, 0.0, 20.0);
+        private IHistogram1D trkChisqProbHist_UnconstrainedMollerVertices = aida.histogram1D("UnconstrainedMollerVertices/Moller Track Chisq Prob", 100, 0.0, 1.0);
+        private IHistogram1D vtxXHist_UnconstrainedMollerVertices = aida.histogram1D("UnconstrainedMollerVertices/Moller Vertex x", 200, -2.5, 2.5);
+        private IHistogram1D vtxYHist_UnconstrainedMollerVertices = aida.histogram1D("UnconstrainedMollerVertices/Moller Vertex y", 200, -1.0, 1.0);
+        private IHistogram1D vtxZHist_UnconstrainedMollerVertices = aida.histogram1D("UnconstrainedMollerVertices/Moller Vertex z", 200, -20.0, 20.0);
+        private IHistogram1D vtxChisqHist_UnconstrainedMollerVertices = aida.histogram1D("UnconstrainedMollerVertices/Moller Vertex Chisq", 100, 0.0, 100.0);
 
-        String[] histoNames = ref.listObjectNames(".", true);
-        String[] histoTypes = ref.listObjectTypes(".", true);
-        System.out.println("comparing " + histoNames.length + " managed objects");
-        double tolerance = 5E-3;
-        for (int i = 0; i < histoNames.length; ++i) {
-            String histoName = histoNames[i];
-            if (histoTypes[i].equals("IHistogram1D")) {
-                System.out.println("Checking entries, means and rms for " + histoName);
-                IHistogram1D h1_r = (IHistogram1D) ref.find(histoName);
-                IHistogram1D h1_t = (IHistogram1D) tst.find(histoName);
-                //System.out.println("           Found              Expected");
-                System.out.println("Entries: "+h1_r.entries()+" "+ h1_t.entries());
-                System.out.println("Mean: "+h1_r.mean()+" "+ h1_t.mean());
-                System.out.println("RMS "+h1_r.rms()+" "+ h1_t.rms());
-            }
+        //2D
+        private IHistogram2D p1vsp2Hist_UnconstrainedMollerVertices = aida.histogram2D("UnconstrainedMollerVertices/Moller p1 vs p2", 200, 0.25, 1.75, 200, 0.25, 1.75);
+        private IHistogram2D ptopvspbotHist_UnconstrainedMollerVertices = aida.histogram2D("UnconstrainedMollerVertices/Moller p top vs p bottom", 200, 0.25, 1.75, 200, 0.25, 1.75);
+        private IHistogram2D theta1vstheta2Hist_UnconstrainedMollerVertices = aida.histogram2D("UnconstrainedMollerVertices/Moller theta1 vs theta2", 100, 0.01, 0.05, 100, 0.01, 0.05);
+        private IHistogram2D pvsthetaHist_UnconstrainedMollerVertices = aida.histogram2D("UnconstrainedMollerVertices/Moller p vs theta", 100, 0.25, 1.75, 100, 0.01, 0.05);
+        private IHistogram2D xvsyHist_UnconstrainedMollerVertices = aida.histogram2D("UnconstrainedMollerVertices/Moller vertex X vs Y", 250, -2.5, 2.5, 100, -1.0, 1.0);
+
+        //
+        private IHistogram1D invMassHist_BeamspotConstrainedMollerVertices = aida.histogram1D("BeamspotConstrainedMollerVertices/Moller Invariant Mass", 200, 0., 0.1);
+        private IHistogram1D pHist_BeamspotConstrainedMollerVertices = aida.histogram1D("BeamspotConstrainedMollerVertices/Moller Momentum", 200, 0., 3.0);
+        private IHistogram1D pxHist_BeamspotConstrainedMollerVertices = aida.histogram1D("BeamspotConstrainedMollerVertices/Moller x Momentum", 200, -0.01, 0.01);
+        private IHistogram1D pyHist_BeamspotConstrainedMollerVertices = aida.histogram1D("BeamspotConstrainedMollerVertices/Moller y Momentum", 200, -0.01, 0.01);
+        private IHistogram1D pzHist_BeamspotConstrainedMollerVertices = aida.histogram1D("BeamspotConstrainedMollerVertices/Moller z Momentum", 200, 0., 3.0);
+        private IHistogram1D trkpHist_BeamspotConstrainedMollerVertices = aida.histogram1D("BeamspotConstrainedMollerVertices/Moller Track Momentum", 100, 0.25, 1.75);
+        private IHistogram1D trkptopHist_BeamspotConstrainedMollerVertices = aida.histogram1D("BeamspotConstrainedMollerVertices/Moller Top Track Momentum", 100, 0.25, 1.75);
+        private IHistogram1D trkpbotHist_BeamspotConstrainedMollerVertices = aida.histogram1D("BeamspotConstrainedMollerVertices/Moller Bottom Track Momentum", 100, 0.25, 1.75);
+        private IHistogram1D trkNhitsHist_BeamspotConstrainedMollerVertices = aida.histogram1D("BeamspotConstrainedMollerVertices/Moller Track Number of Hits", 7, -0.5, 6.5);
+        private IHistogram1D trkChisqHist_BeamspotConstrainedMollerVertices = aida.histogram1D("BeamspotConstrainedMollerVertices/Moller Track Chisq per DoF", 100, 0.0, 20.0);
+        private IHistogram1D trkChisqProbHist_BeamspotConstrainedMollerVertices = aida.histogram1D("BeamspotConstrainedMollerVertices/Moller Track Chisq Prob", 100, 0.0, 1.0);
+        private IHistogram1D vtxXHist_BeamspotConstrainedMollerVertices = aida.histogram1D("BeamspotConstrainedMollerVertices/Moller Vertex x", 200, -2.5, 2.5);
+        private IHistogram1D vtxYHist_BeamspotConstrainedMollerVertices = aida.histogram1D("BeamspotConstrainedMollerVertices/Moller Vertex y", 200, -1.0, 1.0);
+        private IHistogram1D vtxZHist_BeamspotConstrainedMollerVertices = aida.histogram1D("BeamspotConstrainedMollerVertices/Moller Vertex z", 200, -20.0, 20.0);
+        private IHistogram1D vtxChisqHist_BeamspotConstrainedMollerVertices = aida.histogram1D("BeamspotConstrainedMollerVertices/Moller Vertex Chisq", 100, 0.0, 100.0);
+
+        //2D
+        private IHistogram2D p1vsp2Hist_BeamspotConstrainedMollerVertices = aida.histogram2D("BeamspotConstrainedMollerVertices/Moller p1 vs p2", 200, 0.25, 1.75, 200, 0.25, 1.75);
+        private IHistogram2D ptopvspbotHist_BeamspotConstrainedMollerVertices = aida.histogram2D("BeamspotConstrainedMollerVertices/Moller p top vs p bottom", 200, 0.25, 1.75, 200, 0.25, 1.75);
+        private IHistogram2D theta1vstheta2Hist_BeamspotConstrainedMollerVertices = aida.histogram2D("BeamspotConstrainedMollerVertices/Moller theta1 vs theta2", 100, 0.01, 0.05, 100, 0.01, 0.05);
+        private IHistogram2D pvsthetaHist_BeamspotConstrainedMollerVertices = aida.histogram2D("BeamspotConstrainedMollerVertices/Moller p vs theta", 100, 0.25, 1.75, 100, 0.01, 0.05);
+        private IHistogram2D xvsyHist_BeamspotConstrainedMollerVertices = aida.histogram2D("BeamspotConstrainedMollerVertices/Moller vertex X vs Y", 250, -2.5, 2.5, 100, -1.0, 1.0);
+
+    //
+        private IHistogram1D invMassHist_TargetConstrainedMollerVertices = aida.histogram1D("TargetConstrainedMollerVertices/Moller Invariant Mass", 200, 0., 0.1);
+        private IHistogram1D pHist_TargetConstrainedMollerVertices = aida.histogram1D("TargetConstrainedMollerVertices/Moller Momentum", 200, 0., 3.0);
+        private IHistogram1D pxHist_TargetConstrainedMollerVertices = aida.histogram1D("TargetConstrainedMollerVertices/Moller x Momentum", 200, -0.01, 0.01);
+        private IHistogram1D pyHist_TargetConstrainedMollerVertices = aida.histogram1D("TargetConstrainedMollerVertices/Moller y Momentum", 200, -0.01, 0.01);
+        private IHistogram1D pzHist_TargetConstrainedMollerVertices = aida.histogram1D("TargetConstrainedMollerVertices/Moller z Momentum", 200, 0., 3.0);
+        private IHistogram1D trkpHist_TargetConstrainedMollerVertices = aida.histogram1D("TargetConstrainedMollerVertices/Moller Track Momentum", 100, 0.25, 1.75);
+        private IHistogram1D trkptopHist_TargetConstrainedMollerVertices = aida.histogram1D("TargetConstrainedMollerVertices/Moller Top Track Momentum", 100, 0.25, 1.75);
+        private IHistogram1D trkpbotHist_TargetConstrainedMollerVertices = aida.histogram1D("TargetConstrainedMollerVertices/Moller Bottom Track Momentum", 100, 0.25, 1.75);
+        private IHistogram1D trkNhitsHist_TargetConstrainedMollerVertices = aida.histogram1D("TargetConstrainedMollerVertices/Moller Track Number of Hits", 7, -0.5, 6.5);
+        private IHistogram1D trkChisqHist_TargetConstrainedMollerVertices = aida.histogram1D("TargetConstrainedMollerVertices/Moller Track Chisq per DoF", 100, 0.0, 20.0);
+        private IHistogram1D trkChisqProbHist_TargetConstrainedMollerVertices = aida.histogram1D("TargetConstrainedMollerVertices/Moller Track Chisq Prob", 100, 0.0, 1.0);
+        private IHistogram1D vtxXHist_TargetConstrainedMollerVertices = aida.histogram1D("TargetConstrainedMollerVertices/Moller Vertex x", 200, -2.5, 2.5);
+        private IHistogram1D vtxYHist_TargetConstrainedMollerVertices = aida.histogram1D("TargetConstrainedMollerVertices/Moller Vertex y", 200, -1.0, 1.0);
+        private IHistogram1D vtxZHist_TargetConstrainedMollerVertices = aida.histogram1D("TargetConstrainedMollerVertices/Moller Vertex z", 200, -20.0, 20.0);
+        private IHistogram1D vtxChisqHist_TargetConstrainedMollerVertices = aida.histogram1D("TargetConstrainedMollerVertices/Moller Vertex Chisq", 100, 0.0, 100.0);
+
+        //2D
+        private IHistogram2D p1vsp2Hist_TargetConstrainedMollerVertices = aida.histogram2D("TargetConstrainedMollerVertices/Moller p1 vs p2", 200, 0.25, 1.75, 200, 0.25, 1.75);
+        private IHistogram2D ptopvspbotHist_TargetConstrainedMollerVertices = aida.histogram2D("TargetConstrainedMollerVertices/Moller p top vs p bottom", 200, 0.25, 1.75, 200, 0.25, 1.75);
+        private IHistogram2D theta1vstheta2Hist_TargetConstrainedMollerVertices = aida.histogram2D("TargetConstrainedMollerVertices/Moller theta1 vs theta2", 100, 0.01, 0.05, 100, 0.01, 0.05);
+        private IHistogram2D pvsthetaHist_TargetConstrainedMollerVertices = aida.histogram2D("TargetConstrainedMollerVertices/Moller p vs theta", 100, 0.25, 1.75, 100, 0.01, 0.05);
+        private IHistogram2D xvsyHist_TargetConstrainedMollerVertices = aida.histogram2D("TargetConstrainedMollerVertices/Moller vertex X vs Y", 250, -2.5, 2.5, 100, -1.0, 1.0);
+
+        protected void detectorChanged(Detector detector) {
+            beamAxisRotation.setActiveEuler(Math.PI / 2, -0.0305, -Math.PI / 2);
         }
-        for (int i = 0; i < histoNames.length; ++i) {
-            String histoName = histoNames[i];
-            if (histoTypes[i].equals("IHistogram1D")) {
-                System.out.println("checking entries, means and rms for " + histoName);
-                IHistogram1D h1_r = (IHistogram1D) ref.find(histoName);
-                IHistogram1D h1_t = (IHistogram1D) tst.find(histoName);
-                assertEquals(h1_r.entries(), h1_t.entries());
-                assertEquals(h1_r.mean(), h1_t.mean(), tolerance * abs(h1_r.mean()));
-                assertEquals(h1_r.rms(), h1_t.rms(), tolerance * abs(h1_r.rms()));
-            }
+
+        protected void process(EventHeader event) {
+            super.process(event);
+            for (String vertexCollectionName : vertexCollectionNames) {
+
+                List<Vertex> vertices = event.get(Vertex.class, vertexCollectionName);
+                for (Vertex v : vertices) {
+                    aida.tree().cd("/");
+                    aida.tree().mkdirs(vertexCollectionName);
+                    aida.tree().cd(vertexCollectionName);
+                    ReconstructedParticle rp = v.getAssociatedParticle();
+                    int type = rp.getType();
+                    boolean isGbl = TrackType.isGBL(type);
+                    // require GBL tracks in vertex
+                    if (isGbl) {
+                        List<ReconstructedParticle> parts = rp.getParticles();
+                        ReconstructedParticle rp1 = parts.get(0);
+                        ReconstructedParticle rp2 = parts.get(1);
+                        // basic sanity check here, remove full energy electrons (fee)
+                        if (rp1.getMomentum().magnitude() > 1.5 * _beamEnergy || rp2.getMomentum().magnitude() > 1.5 * _beamEnergy) {
+                            continue;
+                        }
+                        // require both reconstructed particles to have a track and a cluster
+                        if (rp1.getClusters().size() != 1) {
+                            continue;
+                        }
+                        if (rp2.getClusters().size() != 1) {
+                            continue;
+                        }
+                        if (rp1.getTracks().size() != 1) {
+                            continue;
+                        }
+                        if (rp2.getTracks().size() != 1) {
+                            continue;
+                        }
+                        Track t1 = rp1.getTracks().get(0);
+                        Track t2 = rp2.getTracks().get(0);
+                        Cluster c1 = rp1.getClusters().get(0);
+                        Cluster c2 = rp2.getClusters().get(0);
+                        double deltaT = ClusterUtilities.getSeedHitTime(c1) - ClusterUtilities.getSeedHitTime(c2);
+                        // require cluster times to be coincident within 2 ns
+                        if (abs(deltaT) > 2.0) {
+                            continue;
+                        }
+                        // require momentum sum to equal beam energy +-
+                        double psum = rp1.getMomentum().magnitude() + rp2.getMomentum().magnitude();
+                        if (psum < psumMin || psum > psumMax) {
+                            continue;
+                        }
+                        //rotate into physiscs frame of reference
+                        Hep3Vector rprot = VecOp.mult(beamAxisRotation, rp.getMomentum());
+                        Hep3Vector p1rot = VecOp.mult(beamAxisRotation, rp1.getMomentum());
+                        Hep3Vector p2rot = VecOp.mult(beamAxisRotation, rp2.getMomentum());
+                        double theta1 = Math.acos(p1rot.z() / p1rot.magnitude());
+                        double theta2 = Math.acos(p2rot.z() / p2rot.magnitude());
+                        double thetasum = theta1 + theta2;
+                        // cut on thetasum
+                        if (thetasum > _thetasumCut) {
+                            continue;
+                        }
+                        // cut on Moller pX
+                        if (abs(rprot.x()) > 0.01) {
+                            continue;
+                        }
+                        // cut on Moller pY
+                        if (abs(rp.getMomentum().y()) > .01) {
+                            continue;
+                        }
+                        double t1ChisqNdf = t1.getChi2() / t1.getNDF();
+                        double t2ChisqNdf = t2.getChi2() / t2.getNDF();
+
+                        double t1ChisqProb = ChisqProb.gammp(t1.getNDF(), t1.getChi2());
+                        double t2ChisqProb = ChisqProb.gammp(t2.getNDF(), t2.getChi2());
+                        // used to cut on prob < 0.995, which corresponds to roughly 3.4
+                        // change this to a cut on chi-squared/dof which people are more familiar with.
+                        // Omar currently cuts on chi-squared <40(!), irrespective of 5 or 6 hit tracks
+                        // let's start at chisq/dof of 8
+                        if (t1ChisqNdf > _trackChi2NdfCut) {//(t1ChisqProb > 0.995) {
+                            continue;
+                        }
+                        if (t2ChisqNdf > _trackChi2NdfCut) {//(t2ChisqProb > 0.995) {
+                            continue;
+                        }
+                        // all cuts passed, let's fill some histograms
+                        Hep3Vector pos = v.getPosition();
+                        double p1 = rp1.getMomentum().magnitude();
+                        double p2 = rp2.getMomentum().magnitude();
+                        if (vertexCollectionName.equals("UnconstrainedMollerVertices")) {
+                            invMassHist_UnconstrainedMollerVertices.fill(rp.getMass());
+                            pHist_UnconstrainedMollerVertices.fill(rp.getMomentum().magnitude());
+                            pxHist_UnconstrainedMollerVertices.fill(rprot.x());
+                            pyHist_UnconstrainedMollerVertices.fill(rprot.y());
+                            pzHist_UnconstrainedMollerVertices.fill(rprot.z());
+                            trkpHist_UnconstrainedMollerVertices.fill(p1);
+                            trkpHist_UnconstrainedMollerVertices.fill(p2);
+                            if (isTopTrack(t1)) {
+                                trkptopHist_UnconstrainedMollerVertices.fill(p1);
+                                ptopvspbotHist_UnconstrainedMollerVertices.fill(p1, p2);
+                            } else {
+                                trkpbotHist_UnconstrainedMollerVertices.fill(p1);
+                            }
+                            if (isTopTrack(t2)) {
+                                trkptopHist_UnconstrainedMollerVertices.fill(p2);
+                            } else {
+                                trkpbotHist_UnconstrainedMollerVertices.fill(p2);
+                            }
+                            trkNhitsHist_UnconstrainedMollerVertices.fill(t1.getTrackerHits().size());
+                            trkNhitsHist_UnconstrainedMollerVertices.fill(t2.getTrackerHits().size());
+                            trkChisqHist_UnconstrainedMollerVertices.fill(t1.getChi2() / t1.getNDF());
+                            trkChisqHist_UnconstrainedMollerVertices.fill(t2.getChi2() / t2.getNDF());
+                            trkChisqProbHist_UnconstrainedMollerVertices.fill(t1ChisqProb);
+                            trkChisqProbHist_UnconstrainedMollerVertices.fill(t2ChisqProb);
+                            vtxXHist_UnconstrainedMollerVertices.fill(pos.x());
+                            vtxYHist_UnconstrainedMollerVertices.fill(pos.y());
+                            vtxZHist_UnconstrainedMollerVertices.fill(pos.z());
+                            vtxChisqHist_UnconstrainedMollerVertices.fill(v.getChi2());
+
+                            p1vsp2Hist_UnconstrainedMollerVertices.fill(p1, p2);
+                            theta1vstheta2Hist_UnconstrainedMollerVertices.fill(theta1, theta2);
+                            pvsthetaHist_UnconstrainedMollerVertices.fill(p1, theta1);
+                            pvsthetaHist_UnconstrainedMollerVertices.fill(p2, theta2);
+                            xvsyHist_UnconstrainedMollerVertices.fill(pos.x(), pos.y());
+                        }
+                        if (vertexCollectionName.equals("BeamspotConstrainedMollerVertices")) {
+                            invMassHist_BeamspotConstrainedMollerVertices.fill(rp.getMass());
+                            pHist_BeamspotConstrainedMollerVertices.fill(rp.getMomentum().magnitude());
+                            pxHist_BeamspotConstrainedMollerVertices.fill(rprot.x());
+                            pyHist_BeamspotConstrainedMollerVertices.fill(rprot.y());
+                            pzHist_BeamspotConstrainedMollerVertices.fill(rprot.z());
+                            trkpHist_BeamspotConstrainedMollerVertices.fill(p1);
+                            trkpHist_BeamspotConstrainedMollerVertices.fill(p2);
+                            if (isTopTrack(t1)) {
+                                trkptopHist_BeamspotConstrainedMollerVertices.fill(p1);
+                                ptopvspbotHist_BeamspotConstrainedMollerVertices.fill(p1, p2);
+                            } else {
+                                trkpbotHist_BeamspotConstrainedMollerVertices.fill(p1);
+                            }
+                            if (isTopTrack(t2)) {
+                                trkptopHist_BeamspotConstrainedMollerVertices.fill(p2);
+                            } else {
+                                trkpbotHist_BeamspotConstrainedMollerVertices.fill(p2);
+                            }
+                            trkNhitsHist_BeamspotConstrainedMollerVertices.fill(t1.getTrackerHits().size());
+                            trkNhitsHist_BeamspotConstrainedMollerVertices.fill(t2.getTrackerHits().size());
+                            trkChisqHist_BeamspotConstrainedMollerVertices.fill(t1.getChi2() / t1.getNDF());
+                            trkChisqHist_BeamspotConstrainedMollerVertices.fill(t2.getChi2() / t2.getNDF());
+                            trkChisqProbHist_BeamspotConstrainedMollerVertices.fill(t1ChisqProb);
+                            trkChisqProbHist_BeamspotConstrainedMollerVertices.fill(t2ChisqProb);
+                            vtxXHist_BeamspotConstrainedMollerVertices.fill(pos.x());
+                            vtxYHist_BeamspotConstrainedMollerVertices.fill(pos.y());
+                            vtxZHist_BeamspotConstrainedMollerVertices.fill(pos.z());
+                            vtxChisqHist_BeamspotConstrainedMollerVertices.fill(v.getChi2());
+
+                            p1vsp2Hist_BeamspotConstrainedMollerVertices.fill(p1, p2);
+                            theta1vstheta2Hist_BeamspotConstrainedMollerVertices.fill(theta1, theta2);
+                            pvsthetaHist_BeamspotConstrainedMollerVertices.fill(p1, theta1);
+                            pvsthetaHist_BeamspotConstrainedMollerVertices.fill(p2, theta2);
+                            xvsyHist_BeamspotConstrainedMollerVertices.fill(pos.x(), pos.y());
+                        }
+                        if (vertexCollectionName.equals("TargetConstrainedMollerVertices")) {
+                            invMassHist_TargetConstrainedMollerVertices.fill(rp.getMass());
+                            pHist_TargetConstrainedMollerVertices.fill(rp.getMomentum().magnitude());
+                            pxHist_TargetConstrainedMollerVertices.fill(rprot.x());
+                            pyHist_TargetConstrainedMollerVertices.fill(rprot.y());
+                            pzHist_TargetConstrainedMollerVertices.fill(rprot.z());
+                            trkpHist_TargetConstrainedMollerVertices.fill(p1);
+                            trkpHist_TargetConstrainedMollerVertices.fill(p2);
+                            if (isTopTrack(t1)) {
+                                trkptopHist_TargetConstrainedMollerVertices.fill(p1);
+                                ptopvspbotHist_TargetConstrainedMollerVertices.fill(p1, p2);
+                            } else {
+                                trkpbotHist_TargetConstrainedMollerVertices.fill(p1);
+                            }
+                            if (isTopTrack(t2)) {
+                                trkptopHist_TargetConstrainedMollerVertices.fill(p2);
+                            } else {
+                                trkpbotHist_TargetConstrainedMollerVertices.fill(p2);
+                            }
+                            trkNhitsHist_TargetConstrainedMollerVertices.fill(t1.getTrackerHits().size());
+                            trkNhitsHist_TargetConstrainedMollerVertices.fill(t2.getTrackerHits().size());
+                            trkChisqHist_TargetConstrainedMollerVertices.fill(t1.getChi2() / t1.getNDF());
+                            trkChisqHist_TargetConstrainedMollerVertices.fill(t2.getChi2() / t2.getNDF());
+                            trkChisqProbHist_TargetConstrainedMollerVertices.fill(t1ChisqProb);
+                            trkChisqProbHist_TargetConstrainedMollerVertices.fill(t2ChisqProb);
+                            vtxXHist_TargetConstrainedMollerVertices.fill(pos.x());
+                            vtxYHist_TargetConstrainedMollerVertices.fill(pos.y());
+                            vtxZHist_TargetConstrainedMollerVertices.fill(pos.z());
+                            vtxChisqHist_TargetConstrainedMollerVertices.fill(v.getChi2());
+
+                            p1vsp2Hist_TargetConstrainedMollerVertices.fill(p1, p2);
+                            theta1vstheta2Hist_TargetConstrainedMollerVertices.fill(theta1, theta2);
+                            pvsthetaHist_TargetConstrainedMollerVertices.fill(p1, theta1);
+                            pvsthetaHist_TargetConstrainedMollerVertices.fill(p2, theta2);
+                            xvsyHist_TargetConstrainedMollerVertices.fill(pos.x(), pos.y());
+                        }
+                    } //Loop over GBL-vertices
+                } // Loop over vertices
+            } // loop over various vertex collections
         }
     }
 }
